pyproject.toml
[project]
name = "amlic"
version = "0.1.0"
description = "Adaptive LLM Inference via Prefill/Decode Disaggregation — Columbia COMS 6998"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "openai>=1.10",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "transformers>=4.40",
    "tokenizers>=0.19",
    "streamlit>=1.32",
    "pypdf>=4.0",
    "sentence-transformers>=2.5",
    "faiss-cpu>=1.7",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "tqdm>=4.66",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]
